Explicit define uint -> uint256

+++ contracts/hPools/HPoolManager.sol
@@ -113,7 +113,7 @@ contract HPoolManager is PausableUpgradeable, HordMiddleware {
     /**
      * @notice          Internal function to handle safe transferring of ETH.
      */
-    function safeTransferETH(address to, uint value) internal {
+    function safeTransferETH(address to, uint256 value) internal {
         (bool success,) = to.call{value:value}(new bytes(0));
         require(success, 'TransferHelper: ETH_TRANSFER_FAILED');
     }
